requirements.txt
# Python dependencies for AppImage Re-Signer

# Core GPG library for Python
python-gnupg>=0.5.0

# For better CLI output (optional but recommended)
colorama>=0.4.6
